Add PresentationCompletedView for handling success/error view
